@@ -1,14 +1,27 @@
 package org.httpsknighthacks.knighthacksandroid;
 
 import android.app.ActionBar;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+import android.util.Log;
 import android.view.View;
 import android.view.WindowManager;
+import android.widget.Toast;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.iid.FirebaseInstanceId;
+import com.google.firebase.iid.InstanceIdResult;
+import com.google.firebase.messaging.FirebaseMessaging;
 
 import org.httpsknighthacks.knighthacksandroid.Resources.RequestQueueSingleton;
 
@@ -23,6 +36,8 @@ public class MainActivity extends AppCompatActivity {
     private ArrayList<Class> activityList = new ArrayList<>();
 
     public RequestQueueSingleton mRequestQueueSingleton;
+
+    private String ANNOUNCEMENTS_TOPIC = "ANNOUNCEMENTS";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -48,10 +63,41 @@ public class MainActivity extends AppCompatActivity {
             }
         }
 
+        // Start of firebase notification code.
+
+        // For notifications received in foreground
+        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.O) {
+            NotificationChannel channel =
+                    new NotificationChannel("MyNotifications", "MyNotifications", NotificationManager.IMPORTANCE_DEFAULT);
+
+            NotificationManager manager = getSystemService(NotificationManager.class);
+            manager.createNotificationChannel(channel);
+        }
+
+        subscribeToTopic(ANNOUNCEMENTS_TOPIC);
+
+        // End of firebase code.
+
+
         setContentView(R.layout.activity_main);
 
         mRequestQueueSingleton = new RequestQueueSingleton(getApplicationContext());
         getImageAndTitles();
+    }
+
+    private void subscribeToTopic(final String topic) {
+        FirebaseMessaging.getInstance().subscribeToTopic(topic)
+                .addOnCompleteListener(new OnCompleteListener<Void>() {
+                    @Override
+                    public void onComplete(@NonNull Task<Void> task) {
+                        String msg = "Successful: Subscribed to " + topic;
+                        if (!task.isSuccessful()) {
+                            msg = "Failed";
+                        }
+                        Log.d("Subscription to " + topic, msg);
+                        Toast.makeText(MainActivity.this, msg, Toast.LENGTH_SHORT).show();
+                    }
+                });
     }
 
     // Get images and titles from the database
